[Java] Move test code to where it belongs
